Introduce implementationCompileElements variant
Adds a new variant to all JVM features that exposes the implementation 'compile view' of that feature.
We introduce a new CompileView attribute to differentiate this variant from apiElements.

In effect, apiElements is the API compile view of the feature, while implementationCompileElements is the implementation compile view.
They differ only in that implementationCompileElements includes implementation dependencies while apiElements does not.

The appropriate compatiblity and disambiguation rules have been added to ensure we can consume existing projects without the attribute
and that consumers without the attribute will default to consuming the api compile view without disambiguity.

This change opens the door to tests suites depending on production code via dependency management, so that it does not need to
interact directly with the production code's configurations or source set outputs. Future changes will add public APIs/dependency modifiers
that make it convenient to request the implementation compile view of a feature.